fix: plugin_live to distinct which state_id is event from
