+++ cdfy/lib/cdfy/plugin/state.ex
@@ -87,14 +87,14 @@ defmodule Cdfy.Plugin.State do
 
   @spec dispatch_event(self :: t(), event :: map()) ::
           {:ok, {t(), map() | String.t()}} | {:error, any()}
-  def dispatch_event(self, event) do
+  def dispatch_event(self, %{player_id: player_id} = event) do
     case Caller.handle_event(self.plugin, event) do
       {:ok, ev} ->
         {:ok, {self, ev}}
 
       {:error, e} ->
         Logger.error("error handling event: #{inspect(e)}")
-        %{self | errors: Map.put(self.errors, event["player_id"], e)}
+        %{self | errors: Map.put(self.errors, player_id, e)}
         {:error, e}
     end
   end
